requirements.txt
Brotli==1.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
ffpyplayer==4.5.1
idna==3.6
mutagen==1.47.0
numpy==1.26.3
opencv-python==4.10.0.84
pillow==10.2.0
pycryptodomex==3.20.0
requests==2.32.3
urllib3==2.1.0
websockets==12.0
yt-dlp==2024.7.16
